Changed to doubles for magnification (just to test, it's easy to put back). Seems to work as well as before. Now the rest should be easy. Also fixed the scaling function (pow(2.0, scaleFraction)). Now the zooming is continous.

+++ libs/MAP/GeoPointLayer.cpp
@@ -213,7 +213,7 @@ namespace MAP
 	//-------------------------------------------------------------------------
 	{
 		PixelCoordinate worldPx = item->getCachedPixelLocation( );
-		if ( worldPx.getMagnification() != magnification )
+		if ( (int) worldPx.getMagnification() != magnification )
 		{
 			LonLat lonlat = item->getLocation( );
 			worldPx = lonlat.toPixels( magnification );
